requirements.txt
requests
python-whois
beautifulsoup4
lxml
googlesearch-python
pandas
python-dotenv
joblib
scikit-learn
flask
flask-cors
matplotlib
seaborn
nltk
textblob
spacy
APScheduler
flask-apscheduler
kaggle